Only return a retrieved_temporary_folder FolderData if non-empty
If no files were specified in the 'retrieve_temporary_list' of the
calculations CalcInfo, we should not pass an empty FolderData to
the parse_from_calc call of the Parser. Otherwise the parse_with_retrieved
call will always get a retrieved_temporary_folder, even if it is
empty. By setting retrieved_temporary_folder to None, unless some
files have been specified in the retrieve_temporary_list member
we maintain the original behavior of the execmanager calling the
parser