Properly namespace the logger for sub classes of AbstractNode
The __metaclass__ of the AbstractNode class does some magic to
properly set the logger that is configured for AiiDA for all the
sub classes, however it did not properly prepend it with the aiida
namespace. This means that for sub classes that live outside of
the aiida module, their module path will not match the aiida logger
and the configured handlers and adapters will not be applied.
As a result, for example, in JobCalculations that live in separate
plugins, the DbLogHandler will not work as expected.