Fixes 902
There was a mistake in the logic of the 'if' block of workchains when
loading.  A position varaiable was being used to keep track of how many
times that conditional block was ticked.  This was being upped each
time but also being used to determine which of possible conditional
branches the condition was at e.g.:

if(...)( <-- pos 0

) elif(...) <-- pos 1

) else(

)

Long story short, when the condition was reloaded from a saved state
it was possible that pos was larger than the number of conditions
(usual just one if there's only an if) and it couldn't resume
from where it was when it was saved.