Ensure that corrupt pickle get moved to failed directory
When the loading of a pickle fails, most likely due to it being
corrupted, it might be useful to inspect its state for debugging.
Therefore, instead of deleting it, we move it to the failed pickle
directory. I also make sure that all pickle directories exists
by calling _ensure_directory on the passed paths upon construction
of the Persistence instance