When a pickle is corrupted, attempt to neatly seal the node
If the daemon attempts to load a Process from a pickle that
excepts, because it is corrupted, we should at least try to
load the corresponding node, set it to failed and seal it, such
that any parent processes that might have called it, get a proper
signal that the process they were waiting on has finished, albeit
unsuccessfully. The parent processes will then not linger in a
forever unfinished state, but will be taken out of limbo.

To this end we try to load the node, by getting the pk from the
pickle filename. Note that this is not very robust as if at some
point the format of the pickle name changes and it no longer is
{node_ppk}.pickle than there is no way to retrieve the
corresponding node.