Change the expectation value for failed process in daemon tests
In the previous commit, logic was added to the process loading
code of the daemon that if loading a process from a pickle excepts
the associated WorkCalculation node is properly set to failed and
sealed. The test_create_fail tested that the node of the failed
process was not sealed, which was the old behavior. With the new
changes the node is in fact expected to be sealed, so we change
the expected outcome of the test