Replace outdated, broken and duplicated work examples
Various examples still directly depended on the aiida_quantumespresso
plugin, which has long been removed. These examples rather belong in
the aiida-quantumespresso repository.
Many of the other examples were extremely outdated and used features
of the workflow engine that no longer exists, or as in the case of
async should no longer be used as they will be deprecated. The remaining
examples had a lot of overlap. I have reduced the important feature
demonstrations to a smaller set of examples with short descriptions.