Update the Node store method to use the NodeRepository
Previously when storing a node, the temporary sandbox folder would
be copied to a temporary folder in the repository on the filesystem.
This would now have to change to go through the NodeRepository class.

Initially the copy would be performed first and then the rest of the
storing procedure would be done and if anything would go wrong, the
repository folder would be placed back into the temporary sandbox
folder. However, this is no longer possible, because for the files
to be able to be stored in the permanent Repository, the node instance
in the database needs to exist for the DbFile and DbNodeFiles to be
properly linked to the DbNode instance.