Update the Django tests setup and teardown w.r.t. repository
Before each test, the test repository needs to be created on disk
and the UUID file needs to be written. Afterwards, the corresponding
DbRepository entry needs to be created in the database.

After each test, the DbRepository, DbNodeFile and DbFile data tables
need to be cleaned

Note that currently the tests assume a FileSystem type repository,
however, this need of course not be the case