Added the -q quiet flag to the workflow kill command
When passed, the additional output will not be printed
to the console. This is useful when the command line is
tested and we do not want any output to the screen. With
this change we can pass the -q flag to achieve this goal
and the default behavior of the command will remain unchanged