Add a filter to the aiida.console StreamHandler to suppress logs during testing
Many tests require the logging of messages to test the functionality
but these messages are not actually required to be read in the
test output and rather just clutter the output and obscure the
actually important error messages.

We add a filter to the console StreamHandler of the aiida logger
which based on the value of the TESTING_MODE setting will suppress
the log record or not. This setting is defined in aiida.settings
and is False by default. In the run_tests method of the verdi devel
command we import the setting and set it to True. This will cause
the filter to swallow all non-important log messages.
Doing this in the run_tests method guarantees that this filter is
only active during the call of verdi devel tests