Catch and pass the InvalidOperation for WorkCalculation.kill()
When killing a WorkCalculation, we also call the kill method
on all the children calculations that it called. It may very
well be that those children are already in a terminal state
in which case calling kill will raise an InvalidOperation.
We do not want this to except as the goal of killing the
calculation is achieved, so we catch it and pass