Print stack trace after catching ImportError upon entrypoint.load()
The exception handling of the caught ImportError when calling
entrypoint.load() in the get_plugin() method caused the original
ImportError to be swallowed and just raised the LoadingPluginFailed.
We not also print the stack trace of the ImportError so that the
user knows why the loading of the plugin failed.