Reconfigure the logging for each periodic celery task
This is an attempted solution to fix the broken pipe errors
in the celery worker logging. It probably is not the correct
way of solving it and is merely patching the result of an
incorrect configuration elsewhere.

The logging for AiiDA is configured from a dictionary in
aiida/__init__.py, where for the 'aiida' loggers and all its
children a database and console handler are configured. The
console handler is set to print to stdout. Once this logger is
triggered in the code executed by one of the celery workers of
the daemon process, the pipe to this file descriptor is broken
and an IOError exception is thrown.

The botch in this commit, reconfigures the logging at each task
call, which ensures that the file descriptor of the StreamHandler
is pointing to the correct stdout of the worker.