Improve the solution to fix aiida logging in celery workers tasks
We define a new function aiida.common.log.setup_logging that will
configure the logging from the aiida.LOGGING dictionary. When the
argument 'daemon' is set to True, the logging will be configured
for a daemon process, which is currently being run by celery workers.
The logging configuration is adapted, such that all loggers will
get a RotatingFileHandler handler that writes to the daemon log file
which is the same as the log file of the master celery process that
is launched when the daemon started. This accomplishes that in that
log file both the celery log messages show up, with celery's own
formatting, as well as the log messages from all the loggers
configured by aiida itself.