Remove the daemon_logfile handler from the default logging dict
In the aiida initialization the configure_logging() function will
be called, which will use the aiida.common.log.LOGGIN dictionary
to configure the logger. It defined the 'daemon_logfile'
RotatingFileHandler, with a filename in the daemon log directory
of the AiiDA config directory. The problem is that this directory
does not necessary exist, but when the logging is configured the
file handler setup will break if it doesn't. To solve this we
remove the handler from the default dictionary and only add it
when the configure_logging is called with daemon=True