Remove the async function from aiida.work.run
The async function relied on the ParallelEngine of plum to
launch multiple processes in parallel, however, during testing
we discovered that the orm is not thread safe and race conditions
could lead to unexpected behavior. The future version of plum
that is currently in development for the renewed workflow engine
will not support this for this reason. Therefore we remove this
function here altogether in preparation for the improved workchain
engine.