Allow deprecated set_kpoints_path of KpointsData without cell
In a recent update, various buggy methods of the KpointsData class were
deprecated. The deprecated interface would raise a ValueError if the
set_kpoints_path would be called before a cell had been set, however,
the deprecated methods used to allow for this, as long as the provided
list used explicit reciprocal coordinates and cartesion mode was False.
Always raising when a cell was not defined was therefore breaking backwards
compatibility.

To restore this behavior, we now pass cell=None when it is not set, to the
legacy version of the get_explicit_kpoints_path method. This also required
some minor changes to the helper function analyze_cell

A regression test was added to verify that the old behavior was successfully
restored through these changes