Special check for 'calculation.job.JobCalculation' in _get_calculation_info_row
Calculations of the base class JobCalculation will have the class name string
'calculation.job.JobCalculation'. In the _get_calculation_info_row method
the module will therefore be, after the rsplit on the class name, like

    calculation.job

this does not include a period on the end like non base modules would have.
In this exceptional case, the calculation type that is to be returned should
simply be the 'calculation.job' base module