Serialize and deserialize the workchain context between state saving/loading
The state of a workchain instance needs to be picklable and therefore the
context needs to be properly serialized as it may contain for instance
AiiDA nodes. The ContextMixin was a convenience mixin, but because the
current fix would have to override its load_instance_state method and
the inverse save_instance_state, it becomes mostly pointless. We implement
the required attributes directly in WorkChain class and make sure to
serialize and deserialize the contents of the context before saving
and loading the instance state, respectively