Fix the logic to flatten the nested input dictionary for Process
When setting up the database record for a Process, the dictionary of
inputs, which may be nested, has to be flattened, skipping any inputs
whose corresponding port is marked as non_db, i.e. an input that should
not be stored as a node in the database. The link labels will be created
by concatenating the namespaces with underscores.

Also make sure that the port types of the plum ProcessSpec are overridden
with the versions of aiida core, which have the WithNonDb mixin.