Add tests to verify that nested inputs are all AttributeFrozendicts and thus immutable
Plum is currently not recursing down nested input dictionaries when creating
the parsed input data structure of a Process, which means that only the top
level inputs (reachable through self.inputs from within the Process) are
wrapped in an AttributesFrozendict and therefore immutable. Any inputs in an
input namespace will be bare dictionaries and therefore mutable. This requires
a fix in plum, but two tests have been added that test this behavior. The one
with the namespaced inputs currently fails.