Get rid of vestigial leading underscores for non-storable inputs
The old convention to mark Process inputs as non database storable
was to name them with a leading underscore. Now there is a proper
mixin WithNonDb that is mixed into the InputPort and PortNamespace
classes, which adds the attribute non_db, which can be used to
mark a port as storable or non storable. We remove the underscores
from the predefined Process and Workchain inputs

	* _options
	* _label
	* _description
	* _store_provenance

Workchain implementations and JobProcess submissions by users will
have to be updated to remove these leading underscored from the
input dictionaries as well