Move immutable attribute check back to Node
The check for node attribute immutability, i.e. when a node is stored
its attributes are to be immutable, was moved out of the base Node
class and moved to the Data class and the SealableWithUpdatableAttributes
mixin, in order to allow for Calculation attributes that needed to
be mutable even after storing the node, e.g. the calculation state.
However, the logic was improperly implemented and not applied everywhere
leaving certain nodes with all attributes mutable even after being stored.
To prevent this error from happening in the future, we move the check
back to the Node class. The calculation classes will have to override
this behavior in the future to allow for updatable attributes