Reimplementation of the Sealable concept
The attribues of the Node class become immutable once the node is
stored. However, for some Node subclasses one needs to allow for
certain attributes to be updatable even after storing. An example
is the Calculation class. For it to run, it needs to be stored,
however its running state is stored in an attribute which needs to
be updated despite the underlying node already being stored.

To solve this we introduce the Sealable mixin, which can defined a
tuple of updatable attributes that are mutable even when the node
is stored. It also provides the 'sealed' attribute which when set
even the updatable attributes become immutable