Implement a ProcessBuilder for interactive process input building
The ProcessBuilder will effectively create a FixedFieldsAttributeDict
where the allowed fields correspond to the input ports of the Process
class that it is created for. This will provide autocomplete functionality
to the user with default values already populated. A single additional
method launch() allows the Process to be submitted or ran with the
current state of the attributes dictionary as the inputs dictionary.
This also satisfies that the user does not have to import the run or
submit functions