Implement input port doc strings for ProcessBuilder attributes
The attributes of a ProcessBuilder instance correspond to the names
of the input ports of the process spec that it was created for. The
InputPorts in that ProcessSpec contain all the information that a
user might need, such as a help text, valid types and optional default
values, so it would be good to have that information show up as the
docstring of the attributes of the ProcessBuilder. This is now
implemented and means that the docstring can be viewed interactively

	> PwCalculation = CalculationFactory('quantumespresso.pw')
	> builder = PwCalculation.get_builder()
	> builder.parameters?
	Type:        ProcessBuilderInput
	String form: None
	File:        ~/code/aiida/env/workflows/aiida-core/aiida/work/process_builder.py
	Docstring:   {
		"help": "Use a node that specifies the input parameters for the namelists",
		"name": "parameters",
		"valid_type": "<class 'aiida.orm.data.parameter.ParameterData'>"
	}

This gives the user all the information they need while never having
to leave the interactive shell