Add a JobProcessBuilder specifically for JobProcesses
The JobCalculation class has a method called submit_test that allows
a user to create a fully qualified Calculation node in memory, without
storing it to the database, nor any of its inputs, while creating the
actual calculation directory that would be created on launch. This is
useful to inspect whether the input files that would be generated
based on the current inputs, would be correct, without actually having
to submit it to the scheduler or storing it in the database.

This functionality is now supported through the ProcessBuilder by
subclassing it to the JobProcessBuilder which will be returned for
JobProcesses when calling get_builder() since this functionality only
makes sense for JobProcesses and not Processes in general. The launch
method of this builder takes the optional argument 'test' that is False
by default, but when set to True will create the submit test folder