Ported inline calculations to use workfunction mechanism
Along the way discovered that default values were not checked in inline
calculations for being Data types.  This is now added.

Found bug in workfunctions mechanism that was not using defaults correctly
and therefore before this commit any default values in workfunctions were
at risk of being wrong (by at risk I mean 100% wrong).

Also discovered problem in plumpy when a default value was not passed.  This
was being passed as None and was therefore not being differentiated from an
unspecified value when being validated.