Temporarily change the daemon to 'verdi devel run_daemon'
To have the daemon test script run on Travis, we had to change
the JobCalculations to go properly through the Process level.
However, that means that the old daemon, currently launched by
verdi daemon start, which launches a celery worker, won't be
able to run them at all.

Instead we temporarily replace the celery worker with a subprocess
call to `verdi devel run_daemon` which will run the new daemon
in the background. Note that this is not actually daemonized, but
just runs a DaemonRunner in a separate process. For the tests this
should work for the time being, until issue #1055 is fixed, that
will implement a properly daemonized version of the daemon runner.