Fix a bug in store method of CifData
The policy is that one can .store() on a node as many times as one
wants. If the node is already stored it simply shouldn't do anything.
This was not respected in the overriden store method of the CifData
data class, that always tried to set the md5 attribute which would
except when the node was already stored because then attributes are
immutable