Fix import error and unnecessary backend check in verdi import
The import functionality for files was still using the backend specific
import functions, import_data_dj and import_data_sqla, while the one for
URL's was using the generalized import_data. However, the latter was not
imported. Hence an import error was raised.

The solution is to import import_data function and update the section
for files to use this one as well and not check the backend itself