Reduce number of workchains launched in travis daemon test
In a recent commit the WorkChain class that was used for testing the
daemon on Travis was changed to have arbitrary levels of nesting. The
number of workchains being launched was still 30, leading to roughly 450
workchains in total being launched. This is a bit excessive and was
slowing down the tests. Reducing the number to 8, which will have a
maximum nesting of 8, is still fine.