Reduce the default safe open interval for ssh transport
For some as of yet unknown reason, having the interval for ssh at 30
will fail the daemon tests on Travis, but only for SqlA. The last
job calculation that gets submitted is properly finished, just like
all the other ones, except that its process state attribute is not
updated to be Finished, but rather is stuck in Waiting. This means
that the conditional for all calculations being Terminated is not
satisified (even though all job calculation are FINISHED) and the test
simply times out.