Remove Sealable mixin from AbstractCode and turn 'hidden' into extra
The 'input_plugin', 'append_text' and 'prepend_text' attributes should not
really be updatable as they can really alter the behavior of the code. For
that reason we remove them from the updatable_attributes. The 'hidden'
attribute can be turned into an extra, as it really is more meta-data that
is only pertinent to the user that owns the code. That leaves no updatable
attributes for this class, which means the Sealable mixin can be removed.