Implement the concept of 'finish status' for Calculations
In the new Process paradigm, all Calculations function simply as records
of what happened during execution of the process. The state of the process
is inscribed on the calculation at every state change, so that it can
be reached by proxy through the node. The terminal process state FINISHED
means that it executed nominally and was not killed nor did an exception
occur. However, calculations may want to still distinguish in this case
between a successful and unsuccessful result.

This is now facilitated by the 'finish status'. If a process finishes
nominally, the return value will be interpreted as an exit code and set
as the finish status, a new attribute for the calculation classes.
The properties 'is_finished_ok' and 'is_failed' now allow a user to
distinguish between a successful and an unsuccessful finished process.

This is currently implemented for InlineCalculation and WorkCalculations.
Since inline and workfuntions either return successfully or except, their
finish status will always be 0, so if they are FINISHED they will always
necessarily be FINISHED_OK.

For WorkChains however, this new schema allows users to exit the workchain
with a specific non-zero error code to indicate that it failed, while still
running through nominally. This can be achieved by simply returning a non
zero value from within an outline function. It can also be done in the
outline directly by using the 'return_' instruction. By default this will
pass a zero return value, but a non-zero value can be set by calling it,
e.g. as 'return_(255)'. This will set the finish status of the work
calculation node to 255 and will cause 'is_failed' to return false