Remove vestigial attribute keys from hash_ignored_attributes
With the new unified process state attributes for all Calculations the
old attribute keys that are excluded in the hashing procedure need to
be updated. Since all attributes that are in the _updatable_attributes
tuple will already be considered very often no explicit ignored hash
attributes have to be added.