Add finish status support to JobCalculations
The JobCalculation already has a sort of finish status, the terminal
states of the DbCalcState table, i.e.:

	* SUBMISSIONFAILED
	* RETRIEVALFAILED
	* PARSINGFAILED
	* FAILED

The idea is to map these calc states to a fixed finish status integer
and have the JobProcess layer set it according to the calculation state,
which remains leading. The FAILED status was set when the whole calc
execution was successful, from submission, through retrieval, up until
successful parsing, but the result of the calculation itself was not
considered to be successful, which will vary per calculation. We extend
this paradigm to allow the parsing function to return any non-zero
positive integer to mean a specific calculation failure mode, which can
then be defined on a per calculation plugin basis and used as error
codes. Support for extending the basic enumeration for JobCalculations
the JobCalculationFinishStatus is not yet implemented, and therefore
until that is done, any non standard integers cannot be mapped onto
a error label but instead will show UNKNOWN.