Move base Data types to their own files to respect plugin_type convention
The plugin_type column of the node class is used to store the class that
should be used to load the node from the database. But more importantly
this type string, which is a period separated string is used in querying
to determine class hierarchy. Nodes with similar leading parts of their
plugin_type strings fall under the same class, and when that class is
used in querying, all classes that fall below that umbrella will also be
considered.

Having the base data types Bool, Float, Int and Str not be in their own
separate file, whose module path is what determines the plugin_type, will
break this paradigm. For example, when a user queries for all nodes of
type Int, this will also return all Bool, Float and Str objects as they
all have exactly the same plugin_type string prefix. Moving these classes
back to their own file will restore this original functionality even
though it might not be the best idea to have the plugin_type string be
determined by the module path of the class itself.