Homogenize the output of verdi work list and verdi calculation list
Now that all Calculations are routed through the Process layer and
therefore are guaranteed to share the same information through the
process_state and finish_status attributes, we can update the verdi
commands 'calculation list' and 'work' list to use the same labels,
format and default order. This hopefully creates a more consistent
and therefore more easy to understand output for the user.

Unfortunately since the querying for JobCalculations is relatively
hardcoded in AbstractJobCalculation class, I had to duplicate some
of the logic to format the various data attributes in the two
display commands. Ideally this would be abstracted but that would
take a huge effort.