Abstract storing introspected function data for inline and workfunction
In a recent change, the full source file of a function that was wrapped
by the make_inline decorator to create an InlineCalculation was removed
from the attributes where it used to be stored. However, it is important
to keep it, but to keep database performance we store it in the repository.

The functionality that retrieved the necessary information from the
wrapped function was a standalone function:

	`aiida.utils.calculation.add_source_info`

which hard coded the attribute keys of the information to store. The same
function was used by the implementation of the `workfunction` decorator
which functions similar to the `make_inline` calculation. Recently, the
InlineCalculations were made backwards compatible with the new Process
layer, by funneling the make_inline function through the workfunction
wrapper as well.

It made sennse therefore to centralize the code that stored the information
of the functions wrapped by these decorators. Therefore we introduce a
new mixin `FunctionCalculationMixin` for Calculation classes that are
created by wrapping a plain python function. It provides the method
`store_source_info` that will store the function name, namespace and
starting line number in the source file, in the attributes of the
calculation node. The full source file will be copied into the node's
repository. The mixin defines public property getters to return these
values.