Add profile specific uuid 'UUID' to be used for a.o. RabbitMQ
A config migration already existed to add a uuid to each profile
specifically for the RabbitMQ message queues and exchanges. I
simply replaced it with a generic profile uuid field 'UUID'.
I overwrote the existing config migration because this was added
in this fast moving developers branch 'workflows' and I didn't
want to have another unnecessary intermediate migration.