Fix a bug in List base type when used in verdi shell
Instantiating a new List instance and storing it before calling any method
that would set the internal list would trigger a ModificationNotAllowed
exception. That is to say the following in verdi shell would except:

	from aiida.orm.data import List
	l = List()
	l.store()

This has to do with the fact that the internal list, stored in an attribute,
would not be created until it was accessed. This was done in the _get_list()
However, in this example when calling store, the internal set was not yet
defined as it was stored and afterwards ipython tried to pretty print it
invoking repr() which tried to get the internal list, which didn't exist.

Long story short, this was a bug that only appeared in this way in verdi
shell. For that reason I could not write a test that reproduced the problem.