Move List to its own file and update constructor to allow initialization
The constructor will check for the keyword argument 'list' and if not
passed will set it to an empty list. The parent constructor is then
called with those kwargs, causing the 'set_list' method to be called,
through the 'set' method of the Node class, which is now made public
to make this possible.