Fix bug in FunctionCalculationMixin related to source file storing
The method to store the source file of the wrapped function in the
repository relied on the source file path being returned by the
inspect module to be absolute, as that is a requirement of the
insert_path method of the Folder class. However, for functions
defined for example in a shell or a file that is not in the python
path, a relative path or no path at all will be returned. Instead
we use the Folder class' method create_file_from_filelike and pass
it a file handle opened from the path returned by inspect