Fix the definition of the logger in AbstractNode metaclass generation
The logger of sub classes of AbstractNode was getting the 'aiida.' prefix
prepended to ensure the logging configuration was inherited, but this was
done regardless of the module path. This caused the prefix to be prepended
also if the module path already contained 'aiida.'. To fix this, we only
prepend the prefix if it does not already start with 'aiida.'