Remove the concept of 'process' in profile configuration
With the new circus daemon, every profile can have its very own
daemon, that runs completely isolated and independent of all the
other profile daemons. The original concept of a 'process' in the
profile configuration, which allowed one to set a different profile
for the 'daemon' and the 'verdi' process, no longer makes sense.
Since every profile has its own daemon, a single profile is fully
self contained, for both the verdi and daemon usage.