Move daemonize call after building the arbiter_config
There was a problem with the config.json configuration file, which
led to an exception in one of the profile_daemon.get_endpoint() calls
in 'verdi daemon start'. Because the daemonize call was done before
this, the exception was silently swallowed because the output was
already redirected to the background. By moving the creating of the
arbiter configuration before the daemonize call, the exception will
now be displayed.