Implement the '--reset' option for verdi daemon restart
The default 'verdi daemon restart' will keep the circus daemon process
alive and simply restart the workers. This will thus keep the number
of workers that were running at the invocation of the command.

The '--reset' flag will force the circus daemon process to also be
shutdown and the whole thing to be restarted. This behavior emulates
the calling of 'verdi daemon stop' followed by 'verdi daemon start'
manually. After restart a new circus daemon process will be running
with a single worker