Factor out a private _start_circus entry point for verdi daemon
This allows us to call one single command in a subprocess from both
the 'verdi daemon start' and 'verdi daemon restart --reset' commands.
It also allows us to print a message that we're attempting to start
the daemon and as soon as the subprocess is spawned we can try to
call the client to verify whether it started and notify the user.
This hopefully accomplishes that if the daemon fails to start, it
does not do so completly silently